requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.12.0
